move get_agent_positions to postgres and pass it in to calculate_pnl
